requirements.txt
openai==0.27.4
Pillow==9.5.0
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
pypiwin32==223
comtypes==1.1.14
flake8==6.0.0
autopep8==2.0.2
